THE ONE THE ONLY

+++ src/main/java/com/example/naturelink/repository/IActivityRepository.java
@@ -3,6 +3,8 @@ package com.example.naturelink.repository;
 import com.example.naturelink.entity.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IActivityRepository extends JpaRepository<Activity, Integer> {
+import java.util.List;
+
+public interface IActivityRepository extends JpaRepository<Activity, Long> {
     // Define custom query methods if needed
 }
